photoswipe-example.php
This is some html and js to be used in conjunction with ACF Gallery field to create a slider using photoswipe.  
See also my other file, photoswipe-functions.php, which contains a function to enqueue all the appropriate photoswipe js and css files.
Do remember to actually go get those files and put them in appropriate locations.